# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SAVA API
  description: API del Sistema de Asistencia y Validaciones Académicas (ULEAM)
  version: 1.0.0
  contact:
    name: DIIT - ULEAM
servers:
  - url: /api
    description: Servidor de aplicación
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
...
paths:
...
  /certificado/generar-docx:
    post:
      summary: Generar certificado en formato Word
      security:
        - cookieAuth: []
      responses:
        "200":
          description: DOCX generado
        "401":
          description: No autenticado
...